--- pandas/tests/groupby/methods/test_median.py
@@ -1,38 +1,0 @@
-import numpy as np
-
-import pandas as pd
-from pandas import DataFrame
-import pandas._testing as tm
-
-
-def test_cython_median():
-    arr = np.random.default_rng(2).standard_normal(1000)
-    arr[::2] = np.nan
-    df = DataFrame(arr)
-
-    labels = np.random.default_rng(2).integers(0, 50, size=1000).astype(float)
-    labels[::17] = np.nan
-
-    result = df.groupby(labels).median()
-    msg = "using DataFrameGroupBy.median"
-    with tm.assert_produces_warning(FutureWarning, match=msg):
-        exp = df.groupby(labels).agg(np.nanmedian)
-    tm.assert_frame_equal(result, exp)
-
-    df = DataFrame(np.random.default_rng(2).standard_normal((1000, 5)))
-    msg = "using DataFrameGroupBy.median"
-    with tm.assert_produces_warning(FutureWarning, match=msg):
-        rs = df.groupby(labels).agg(np.median)
-    xp = df.groupby(labels).median()
-    tm.assert_frame_equal(rs, xp)
-
-
-def test_median_empty_bins(observed):
-    df = DataFrame(np.random.default_rng(2).integers(0, 44, 500))
-
-    grps = range(0, 55, 5)
-    bins = pd.cut(df[0], grps)
-
-    result = df.groupby(bins, observed=observed).median()
-    expected = df.groupby(bins, observed=observed).agg(lambda x: x.median())
-    tm.assert_frame_equal(result, expected)
